Fixed error handling
For synchroneous controllers express provides a default error handler.
For asynchroneous controllers we need to catch errors and
pass them to the next middleware (error handler).
The MovieService is synchroneous, so async signatures were removed.
Additionally a custom error handler was added to return a JSON response
in case of internal server errors.